backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
kafka-python
pydantic-settings
langfuse
reportlab
openai
tavily-python
pypdf
